designPatterns/src/test/java/com/sandisiwe/designpaterns/test/structural/ProxyTest.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */

package com.sandisiwe.designpaterns.test.structural;

import com.sandisiwe.designpatterns.structural.proxy.FastThing;
import com.sandisiwe.designpatterns.structural.proxy.Proxy;
import org.testng.Assert;
import org.testng.annotations.AfterClass;
import org.testng.annotations.AfterMethod;
import org.testng.annotations.BeforeClass;
import org.testng.annotations.BeforeMethod;
import org.testng.annotations.Test;

/**
 *
 * @author sandisiwe
 */
public class ProxyTest {
    
    public ProxyTest() {
    }

    // TODO add test methods here.
    // The methods must be annotated with annotation @Test. For example:
    //
    // @Test
    // public void hello() {}
    
    @Test
    public void test() {
        Proxy proxy = new Proxy();
        FastThing fastThing = new FastThing();
        Assert.assertEquals(fastThing.sayHello(), "FastThing says warup mate");
        Assert.assertEquals(proxy.sayHello(), "SlowThing says warup mate");
    }

    @BeforeClass
    public static void setUpClass() throws Exception {
    }

    @AfterClass
    public static void tearDownClass() throws Exception {
    }

    @BeforeMethod
    public void setUpMethod() throws Exception {
    }

    @AfterMethod
    public void tearDownMethod() throws Exception {
    }
}
